# backend/consumption-service/docs/openapi.json
{"openapi":"3.1.0","info":{"title":"HarmoniWatts Consumption Service","description":"Servicio de ingesta y consulta de datos de consumo energético","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/ready":{"get":{"summary":"Readiness Check","description":"Verifica que el servicio esté listo para recibir tráfico","operationId":"readiness_check_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/consumption/current/{household_id}":{"get":{"tags":["consumption"],"summary":"Consumo actual","description":"Obtiene el consumo actual en kW para una vivienda","operationId":"get_current_consumption_api_v1_consumption_current__household_id__get","parameters":[{"name":"household_id","in":"path","required":true,"schema":{"type":"string","description":"ID de la vivienda","title":"Household Id"},"description":"ID de la vivienda"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CurrentConsumptionResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/consumption/daily-total/{household_id}":{"get":{"tags":["consumption"],"summary":"Consumo acumulado del día","description":"Obtiene el consumo total acumulado del día en kWh y su costo","operationId":"get_daily_total_api_v1_consumption_daily_total__household_id__get","parameters":[{"name":"household_id","in":"path","required":true,"schema":{"type":"string","description":"ID de la vivienda","title":"Household Id"},"description":"ID de la vivienda"},{"name":"date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Fecha YYYY-MM-DD","title":"Date"},"description":"Fecha YYYY-MM-DD"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DailyTotalResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/consumption/series/hourly/{household_id}":{"get":{"tags":["consumption"],"summary":"Serie horaria del día","description":"Obtiene el consumo por hora para un día específico","operationId":"get_hourly_series_api_v1_consumption_series_hourly__household_id__get","parameters":[{"name":"household_id","in":"path","required":true,"schema":{"type":"string","description":"ID de la vivienda","title":"Household Id"},"description":"ID de la vivienda"},{"name":"date","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"description":"Fecha YYYY-MM-DD","title":"Date"},"description":"Fecha YYYY-MM-DD"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSeriesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/consumption/series/daily/{household_id}":{"get":{"tags":["consumption"],"summary":"Serie diaria del mes","description":"Obtiene el consumo por día para un mes específico","operationId":"get_daily_series_api_v1_consumption_series_daily__household_id__get","parameters":[{"name":"household_id","in":"path","required":true,"schema":{"type":"string","description":"ID de la vivienda","title":"Household Id"},"description":"ID de la vivienda"},{"name":"year","in":"query","required":true,"schema":{"type":"integer","maximum":2030,"minimum":2020,"description":"Año (ej: 2025)","title":"Year"},"description":"Año (ej: 2025)"},{"name":"month","in":"query","required":true,"schema":{"type":"integer","maximum":12,"minimum":1,"description":"Mes (1-12)","title":"Month"},"description":"Mes (1-12)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSeriesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/consumption/series/monthly/{household_id}":{"get":{"tags":["consumption"],"summary":"Serie mensual del año","description":"Obtiene el consumo por mes para un año específico","operationId":"get_monthly_series_api_v1_consumption_series_monthly__household_id__get","parameters":[{"name":"household_id","in":"path","required":true,"schema":{"type":"string","description":"ID de la vivienda","title":"Household Id"},"description":"ID de la vivienda"},{"name":"year","in":"query","required":true,"schema":{"type":"integer","maximum":2030,"minimum":2020,"description":"Año (ej: 2025)","title":"Year"},"description":"Año (ej: 2025)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSeriesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/consumption/series/yearly/{household_id}":{"get":{"tags":["consumption"],"summary":"Serie anual multi-año","description":"Obtiene el consumo por año para un rango de años","operationId":"get_yearly_series_api_v1_consumption_series_yearly__household_id__get","parameters":[{"name":"household_id","in":"path","required":true,"schema":{"type":"string","description":"ID de la vivienda","title":"Household Id"},"description":"ID de la vivienda"},{"name":"start_year","in":"query","required":true,"schema":{"type":"integer","maximum":2030,"minimum":2020,"description":"Año inicial","title":"Start Year"},"description":"Año inicial"},{"name":"end_year","in":"query","required":true,"schema":{"type":"integer","maximum":2030,"minimum":2020,"description":"Año final","title":"End Year"},"description":"Año final"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TimeSeriesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"CurrentConsumptionResponse":{"properties":{"id_vivienda":{"type":"integer","title":"Id Vivienda"},"current_power_kw":{"type":"number","title":"Current Power Kw"},"last_update":{"type":"string","format":"date-time","title":"Last Update"},"unit":{"type":"string","title":"Unit","default":"kW"}},"type":"object","required":["id_vivienda","current_power_kw","last_update"],"title":"CurrentConsumptionResponse","description":"Respuesta para consumo actual"},"DailyTotalResponse":{"properties":{"id_vivienda":{"type":"integer","title":"Id Vivienda"},"date":{"type":"string","title":"Date"},"total_consumption_kwh":{"type":"number","title":"Total Consumption Kwh"},"total_cost_cop":{"type":"number","title":"Total Cost Cop"},"average_power_kw":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Average Power Kw"},"peak_power_kw":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Peak Power Kw"},"peak_hour":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Peak Hour"},"unit_kwh":{"type":"string","title":"Unit Kwh","default":"kWh"}},"type":"object","required":["id_vivienda","date","total_consumption_kwh","total_cost_cop"],"title":"DailyTotalResponse","description":"Respuesta para consumo acumulado del día"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TimeSeriesResponse":{"properties":{"id_vivienda":{"type":"integer","title":"Id Vivienda"},"granularity":{"type":"string","title":"Granularity"},"start_date":{"type":"string","title":"Start Date"},"end_date":{"type":"string","title":"End Date"},"labels":{"items":{"type":"string"},"type":"array","title":"Labels"},"values_kwh":{"items":{"type":"number"},"type":"array","title":"Values Kwh"},"values_cost":{"anyOf":[{"items":{"type":"number"},"type":"array"},{"type":"null"}],"title":"Values Cost"},"unit":{"type":"string","title":"Unit","default":"kWh"}},"type":"object","required":["id_vivienda","granularity","start_date","end_date","labels","values_kwh"],"title":"TimeSeriesResponse","description":"Respuesta genérica para series temporales"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}